renamed app/ to src/, moved main.py and dbbuilder.py to root directory
