+++ main.py
@@ -1,8 +1,8 @@
 # pylint: disable=import-error
 import sys
 import psycopg2
-from connector import Connector
-from browser import Browser
+from src.connector import Connector
+from src.browser import Browser
 
 if __name__ == "__main__":
     if len(sys.argv) > 1:
