+++ src/data/woodworking.stackexchange.com/Badges.xml
